requirements/requirements.txt
# SHA1:56f04e9aedad7a3b42c70a679bed3175391dd1e2
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   pip-compile-multi
    #   pip-tools
filelock==3.17.0
    # via huggingface-hub
fsspec==2025.2.0
    # via huggingface-hub
huggingface-hub==0.28.1
    # via -r requirements/requirements.in
idna==3.10
    # via requests
packaging==24.2
    # via
    #   build
    #   huggingface-hub
pip-compile-multi==2.7.1
    # via -r requirements/requirements.in
pip-tools==7.4.1
    # via pip-compile-multi
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements/requirements.in
pyyaml==6.0.2
    # via huggingface-hub
requests==2.32.3
    # via huggingface-hub
toposort==1.10
    # via pip-compile-multi
tqdm==4.67.1
    # via huggingface-hub
typing-extensions==4.12.2
    # via huggingface-hub
urllib3==2.3.0
    # via requests
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
